test(resources): assert the tmpfs, driver_opts, build target and orphan sweep (#1278)
Four of the nine bare tests in `build_resources.rs`, all reading state
back rather than trusting the command's return value.

## Two fixtures could not discriminate

| test | why no assertion could have worked |
|---|---|
| `build_with_target_stage` | both stages ran `RUN echo`, which leaves
nothing behind, so an image built from the wrong stage is
indistinguishable from the right one. Each stage now writes its name to
`/stage` |
| `named_volume_with_driver_opts` | without the opts taking effect the
container still gets a working `/cache` — an ordinary local volume — so
the file has to be looked for **on the host** |

## A claim that was prose is now a check

`service_with_expose_deploy_labels_annotations_tmpfs` carried this in a
comment:

> deploy.labels are accepted but, per the Compose Specification, are set
on the service only and must not be applied to the container

That is falsifiable, so it is checked now — alongside the annotation
that *must* reach the container, and the tmpfs that must really be a
tmpfs, read from `/proc/mounts` inside it (no seam needed).

`remove_orphans_removes_container` is the other half of the sibling test
in `lifecycle.rs`. That one pins that a sweep with no orphan present
removes nothing; this one pins that a real orphan goes. Until now, a
sweep that removed nothing at all satisfied both.

## Test plan

Podman 5.7.0, `--test-threads=2`: **17 passed, 0 failed** (73s). `cargo
fmt --check` and clippy clean.

Three mutations, sources restored between rounds and the binary verified
free of markers afterwards:

| mutation | assertion that went red |
|---|---|
| the `&target=` query parameter dropped | `left: "final-stage"`,
`right: "base-stage"` |
| the orphan sweep iterates an empty list | svc1 survives |
| the volume is created without `driver_opts` | `left: ""`, `right:
"bound"` |

**The tmpfs/labels/annotations assertions are not mutation-proved.** The
`deploy.labels` one asserts an *absence*, which needs a mutation that
starts applying them rather than one that stops, and I have not written
it.

## File split

The assertions took the file to 508 code lines, past the 500 hard limit,
so it is cut at its own "Build" section marker: configs, container
options and the driver_opts volume stay, and build, networks, the
long-form refs and the orphan sweep move to `build_images.rs`. 154 and
355 code lines.

I measured that after pushing rather than before, for the second time
today. The gate caught it both times, which is the argument for having
it in CI — but it is one command and I should be running it.

Refs #1180

---------
